Rename pysph.base.opencl to pysph.base.device_helper
