+++ pysph/base/gpu_domain_manager.py
@@ -1,8 +1,5 @@
 import numpy as np
-import pyopencl as cl
-import pyopencl.array
-
-from pysph.base.opencl import get_config
+from pysph.cpy.config import get_config
 
 
 class GPUDomainManager(object):
